[chore] 테스트 코드 위치 변경
통일성을 위해서 패키지를 변경하였습니다.
